chore(tools): draft package refactor (#4825)
