+++ tools/metabase_dashboard_templates/metabase_flow/gcp_secrets.py
