+++ tools/metabase_dashboard_templates/pyproject.toml
@@ -15,5 +15,11 @@ dev = [
     "pytest",
 ]
 
+[tool.pytest.ini_options]
+# Put the tool root on sys.path so tests can `import cli` and
+# `import metabase_flow.*` (the package is run/tested from this directory).
+pythonpath = ["."]
+testpaths = ["tests"]
+
 [tool.uv]
 package = false
